First Commit of the Group 3 Rep
Used the work of Mirco, which implements the functionality fine and uses a
nice OOP approach. While the deletion are not correctly supported yet, it is
almost done and already has some tests regarding this.

Comments probably should be similar to answer as far as the model is concerned.

I also added a generated UML diagram as a quick reference.