Searchengine pt1
Wrote a test that does not yet run greenly, as there seems to be a problem
with adding questions to the database. So yeah, there is some work on the
database now. Also there is yet to be a view for the search and a searchbox. I
hope, this is a minor problem.