Search in answers
The answers are taken into account when searching for a term. Sofar, there is
no real integration into the view: a question is shown as match, but not the
answer. As there might be multiple answers contributing to the match, we got
to figure out, which answers are to be shown and how.